requirements.txt
absl-py==0.8.1
astor==0.8.0
Augmentor==0.2.6
colorama==0.4.1
cycler==0.10.0
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.0
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
scipy==1.3.1
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
tqdm==4.36.1
Werkzeug==0.16.0
yattag==1.12.2
